normalize Plant/Project Name and Type first
the "Power Plant Name" and "Project Name" are normalized first, ensuring (1) all records in a given set have the same Power Plant Name, (2) Project Name === Power Plant Name when it's a "Plant", and (3) "Type" is "Plant" or "Project". Then for the rest of the normalization functions, the record["Type"] can be used directly. This is much more reliable than comparing plant & project names, especially when the plant & project names hadn't been normalized!